% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{autocart}
\alias{autocart}
\title{Return the numeric vector of predictions from an autocart tree.}
\usage{
autocart(response, data, locations, alpha)
}
\arguments{
\item{response}{A vector of numeric response values with no NA entries.}

\item{data}{A dataframe for the predictor variables used in the autocart tree.}

\item{locations}{A two-column matrix with coordinates for the observations the predictor dataframe.}

\item{alpha}{A scalar value between 0 and 1 to weight autocorrelation against reduction in variance in the tree splitting. A value of 1 indicates full weighting on measures of autocorrelation.}
}
\value{
A predicted response value for each of the dataframe observations that were used to create the tree.
}
\description{
Return the numeric vector of predictions from an autocart tree.
}
