% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoforest.R
\name{autoforest}
\alias{autoforest}
\title{Create a forest of autocart trees..}
\usage{
autoforest(response, data, locations, alpha, beta, control, numtrees)
}
\arguments{
\item{response}{The response vector that goes along with the dataframe of predictors.}

\item{data}{The dataframe of predictors.}

\item{locations}{A matrix of the locations of the dataframe of predictors.}

\item{alpha}{The percentage of weighting on spatial autocorrelation in the splitting function.}

\item{beta}{The percentage of weighting on spatial compactness in the splitting function.}

\item{control}{A control object from the \code{autocartControl} function that will be used for each tree in the forest.}

\item{numtrees}{The number of autocart trees to create in the forest.}
}
\value{
An object of type "autoforest", which is a list of the autocart trees.
}
\description{
Create a forest of autocart trees..
}
