% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autocart.R
\name{predict.autocart}
\alias{predict.autocart}
\title{Given an autocart model object, predict for new data passed in}
\usage{
\method{predict}{autocart}(model, newdata, spatialNodes = FALSE, p = NULL, pRange = NULL)
}
\arguments{
\item{model}{An autocart object returned from the autocart() function}

\item{newdata}{A dataframe for prediction. If using spatialNodes = TRUE, then
this dataframe must be a SpatialPointsDataFrame.}

\item{spatialNodes}{A boolean indicating whether or not to use a spatial process
at the terminal nodes of the autocart tree}

\item{p}{The power to use in IDW interpolation when using spatial nodes.}

\item{pRange}{A range of powers to use in IDW interpolation when using spatial nodes.}
}
\value{
A numeric vector containing the predicted response value for each of the rows in the passed in dataframe.
}
\description{
Given an autocart model object, predict for new data passed in
}
\examples{
# Load some data for an autocart predict example
snow <- na.omit(read.csv(system.file("extdata", "ut2017_snow.csv", package = "autocart")))
y <- snow$yr50[1:40]
X <- data.frame(snow$ELEVATION, snow$MCMT, snow$PPTWT, snow$HUC)[1:40, ]
locations <- as.matrix(cbind(snow$LONGITUDE, snow$LATITUDE))[1:40, ]

snow_model <- autocart(y, X, locations, 0.30, 0)

# Predict in autocart
new_X <- X[1:10, ]
new_loc <- locations[1:10, ]
autocart_predictions <- predictAutocart(snow_model, new_X)
}
