% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{predictAutocart}
\alias{predictAutocart}
\title{Given an autocart model object, predict for new data passed in}
\usage{
predictAutocart(autocartModel, newdata)
}
\arguments{
\item{autocartModel}{An S3 object of type "autocart" returned from the autocart function}

\item{newdata}{A dataframe with the same amount of columns used to create the autocart model.}
}
\value{
A numeric vector containing the predicted response value for each of the rows in the passed in dataframe.
}
\description{
Given an autocart model object, predict for new data passed in
}
