% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialNodes.R
\name{predictSpatialNodes}
\alias{predictSpatialNodes}
\title{Using a spatial nodes process, predict with new observations}
\usage{
predictSpatialNodes(
  autocartModel,
  newdata,
  newdataCoords,
  resolution,
  optionalCRS = "+proj=utm +zone=12 +units=km",
  distpower = 2
)
}
\arguments{
\item{autocartModel}{The model returned from the \code{autocart} function}

\item{newdata}{The dataframe of predictor values that we wish to make new predictions with}

\item{newdataCoords}{The matrix of locations for the newdata dataframe}

\item{resolution}{When performing IDW interpolation, what resolution (size of cells) should be used? Make sure this lines up intuitively with the optionalCRS argument.}

\item{optionalCRS}{What coordinate reference system would you like to use in the returned predictions?}

\item{distpower}{The power of distance (e.g. distance squared) used in IDW interpolation}
}
\value{
A vector with the predicted values for the passed in dataframe of observations
}
\description{
Using a spatial nodes process, predict with new observations
}
