% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autocart.R
\name{rmae}
\alias{rmae}
\title{Relative mean absolute error}
\usage{
rmae(pred, obs, na.rm = TRUE)
}
\arguments{
\item{pred}{The vector of predictions}

\item{obs}{The actual observed values}

\item{na.rm}{Should NA values be taken out of the vectors?}
}
\value{
The relative mean average error of the two vectors.
}
\description{
Relative mean absolute error
}
\examples{
# Load some data for an autotune example
snow <- na.omit(read.csv(system.file("extdata", "ut2017_snow.csv", package = "autocart")))
y <- snow$yr50
X <- data.frame(snow$ELEVATION, snow$MCMT, snow$PPTWT)
locations <- as.matrix(cbind(snow$LONGITUDE, snow$LATITUDE))

# Find optimal parameters via cross-validation. We'll search through the
# following alpha/beta/bandwidth values:
alphaVec <- c(0.0, 0.33, 0.66, 1.0)
betaVec <- c(0.0, 0.33, 0.66, 1.0)
bandwidthVec <- c(0.0, 0.33, 0.66, 1.0)

# We'll find the optimal values with 5-fold cross validation:
# (Due to the large number of cross-validations and trainings that occur,
# this can take a few minutes.)
myTune <- autotune(y, X, locations, k = 5, alphaVals = alphaVec,
                   betaVals = betaVec, bandwidthVals = bandwidthVec)
# Inspect the results
myTune
}
