% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialNodes.R
\name{spatialNodes}
\alias{spatialNodes}
\title{Create a spatial process at the end of the nodes of an autocart model.}
\usage{
spatialNodes(autocartModel)
}
\arguments{
\item{autocartModel}{An S3 object of type "autocart" returned by the \code{autocart} function.}
}
\value{
The number 3
}
\description{
Create a spatial process at the end of the nodes of an autocart model.
}
