% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{autocart}
\alias{autocart}
\title{Create an autocart model}
\usage{
autocart(response, data, locations, alpha)
}
\arguments{
\item{response}{A vector of numeric response values with no NA entries.}

\item{data}{A dataframe for the predictor variables used in the autocart tree.}

\item{locations}{A two-column matrix with coordinates for the observations the predictor dataframe.}

\item{alpha}{A scalar value between 0 and 1 to weight autocorrelation against reduction in variance in the tree splitting. A value of 1 indicates full weighting on measures of autocorrelation.}
}
\value{
An S3 object of class "autocart".
}
\description{
Create an autocart model
}
