% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autocart.R
\name{autocart}
\alias{autocart}
\title{Create an autocart model}
\usage{
autocart(formula, data, alpha, beta, control = NULL)
}
\arguments{
\item{formula}{An R formula specifying the model.}

\item{data}{A SpatialPointsDataFrame that contains all the information we will be splitting with,
along with the coordinate information attached to the dataframe.}

\item{alpha}{A scalar value between 0 and 1 to weight autocorrelation against reduction in variance in the tree splitting. A value of 1 indicates full weighting on measures of autocorrelation.}

\item{beta}{A scalar value between 0 and 1 to weight the shape of the region in the splitting}

\item{control}{An object of type "autocartControl" returned by the \code{autocartControl} function to control the splitting in the autocart tree.}
}
\value{
An S3 object of class "autocart".
}
\description{
Create an autocart model
}
\examples{
# Load some data for an autocart example
snow <- na.omit(read.csv(system.file("extdata", "ut2017_snow.csv", package = "autocart")))
snow <- snow[1:40, c("yr50", "LONGITUDE", "LATITUDE", "ELEVATION", "MCMT", "PPTWT", "HUC")]
snow <- sp::SpatialPointsDataFrame(coords = snow[, c("LONGITUDE", "LATITUDE")],
                                   data = snow)

# Create an autocart model
snow_model <- autocart(yr50 ~ ELEVATION + MCMT + PPTWT + HUC, data = snow,
                       alpha = 0.30, beta = 0)

}
