% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autocart.R
\name{autocartControl}
\alias{autocartControl}
\title{Create the object used for the controlling of the splits in the autocart model}
\usage{
autocartControl(
  minsplit = 20,
  minbucket = round(minsplit/3),
  maxdepth = 30,
  distpower = 1,
  islonglat = TRUE,
  standardizeloss = TRUE,
  givePredAsFactor = TRUE,
  retainCoords = TRUE
)
}
\arguments{
\item{minsplit}{The minimum observations in a node before a split is attempted}

\item{minbucket}{The minimum number of observations in a terminal node.}

\item{maxdepth}{Set the maximum depth in the final tree. A root node is counted as a height of 0.}

\item{distpower}{The power of inverse distance to use when calculating spatial weights matrix.}

\item{islonglat}{Are the coordinates longitude and latitude coordinates? If TRUE, then use great circle distance calculations}

\item{standardizeloss}{Measures of autocorrelation, size, and reduction in variance carry a different distribution even though the code scales them between 0 and 1. Should they be standardized to be weighted equally?}

\item{givePredAsFactor}{In the returned autocart model, should the prediction vector also be returned as a factor?}

\item{retainCoords}{After creating the autocart model, should the coordinates for each of the predictions be kept in the returned model?}
}
\value{
An object passed in to the \code{autocart} function that controls the splitting
}
\description{
Create the object used for the controlling of the splits in the autocart model
}
