% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autocart.R
\name{autocartControl}
\alias{autocartControl}
\title{Create the object used for the controlling of the splits in the autocart model}
\usage{
autocartControl(
  minsplit = 20,
  minbucket = round(minsplit/3),
  maxdepth = 30,
  maxobsMtxCalc = NULL,
  distpower = 1,
  islonglat = TRUE,
  givePredAsFactor = TRUE,
  retainCoords = TRUE,
  useGearyC = FALSE,
  saddlepointApproximation = FALSE,
  spatialWeightsType = "default",
  customSpatialWeights = NULL,
  spatialBandwidthProportion = 1,
  spatialBandwidth = NULL
)
}
\arguments{
\item{minsplit}{The minimum observations in a node before a split is attempted}

\item{minbucket}{The minimum number of observations in a terminal node.}

\item{maxdepth}{Set the maximum depth in the final tree. A root node is counted as a height of 0.}

\item{maxobsMtxCalc}{Optional maximum number of observations in a node where computationally intensive matrix calculations like autocorrelation and compactness are performed.}

\item{distpower}{The power of inverse distance to use when calculating spatial weights matrix.}

\item{islonglat}{Are the coordinates longitude and latitude coordinates? If TRUE, then use great circle distance calculations}

\item{givePredAsFactor}{In the returned autocart model, should the prediction vector also be returned as a factor?}

\item{retainCoords}{After creating the autocart model, should the coordinates for each of the predictions be kept in the returned model?}

\item{useGearyC}{Should autocart use Geary's C instead of Moran's I in the splitting function?}

\item{saddlepointApproximation}{Use the saddlepoint approximation to Moran's I. On large datasets this may reduce the compute time by a large amount.}

\item{spatialWeightsType}{What type of spatial weighting should be used when calculating spatial autocorrelation? Options are "default" or "gaussian".}

\item{customSpatialWeights}{Use this parameter to pass in an optional spatial weights matrix for use in autocorrelation calculations. Must have nrow and ncol equal to rows in training dataframe.}

\item{spatialBandwidthProportion}{What percentage of the maximum pairwise distances should be considered the maximum distance for spatial influence? Cannot be simultaneously set with \code{spatialBandwidth}}

\item{spatialBandwidth}{What is the maximum distance where spatial influence can be assumed? Cannot be simultaneously set with \code{spatialBandwidthProportion}.}
}
\value{
An object passed in to the \code{autocart} function that controls the splitting.
}
\description{
Create the object used for the controlling of the splits in the autocart model
}
