% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoforest.R
\name{autoforest}
\alias{autoforest}
\title{Create a forest of autocart trees.}
\usage{
autoforest(
  formula,
  data,
  alpha,
  beta,
  control = autocartControl(),
  numtrees = 50,
  mtry = NULL
)
}
\arguments{
\item{formula}{An R formula specifying the model.}

\item{data}{A SpatialPointsDataFrame that contains all the information we will be splitting with,
along with the coordinate information attached to the dataframe.}

\item{alpha}{A scalar value between 0 and 1 to weight autocorrelation against reduction in variance in the tree splitting. A value of 1 indicates full weighting on measures of autocorrelation.}

\item{beta}{A scalar value between 0 and 1 to weight the shape of the region in the splitting}

\item{control}{An object of type "autocartControl" returned by the \code{autocartControl} function to control the splitting in the autocart tree.}

\item{numtrees}{The number of autocart trees to create in the forest.}

\item{mtry}{The number of variables to subset at each node of the splitting in the trees. By default, this will be 1/3 of the features.}
}
\value{
An object of type "autoforest", which is a list of the autocart trees.
}
\description{
Create a forest of autocart trees.
}
\examples{
# Load some data for an autoforest example
snow <- na.omit(read.csv(system.file("extdata", "ut2017_snow.csv", package = "autocart")))
snow <- snow[1:40, c("yr50", "LONGITUDE", "LATITUDE", "ELEVATION", "MCMT", "PPTWT", "HUC")]
snow <- sp::SpatialPointsDataFrame(coords = snow[, c("LONGITUDE", "LATITUDE")],
                                   data = snow)

# Create an autoforest model with 5 trees
snow_model <- autoforest(yr50 ~ ELEVATION + MCMT + PPTWT + HUC, data = snow,
                       alpha = 0.30, beta = 0, numtrees = 5)
}
