% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autocart.R
\name{predict.autocart}
\alias{predict.autocart}
\title{Given an autocart model object, predict for new data passed in}
\usage{
\method{predict}{autocart}(object, newdata, spatialNodes = FALSE, p = NULL, pRange = NULL, ...)
}
\arguments{
\item{object}{An autocart object returned from the autocart() function}

\item{newdata}{A dataframe for prediction. If using spatialNodes = TRUE, then
this dataframe must be a SpatialPointsDataFrame.}

\item{spatialNodes}{A boolean indicating whether or not to use a spatial process
at the terminal nodes of the autocart tree}

\item{p}{The power to use in IDW interpolation when using spatial nodes.}

\item{pRange}{A range of powers to use in IDW interpolation when using spatial nodes.}

\item{...}{Extra parameters for predict}
}
\value{
A numeric vector containing the predicted response value for each of the rows in the passed in dataframe.
}
\description{
Given an autocart model object, predict for new data passed in
}
\examples{
# Load some data for an autocart example
snow <- na.omit(read.csv(system.file("extdata", "ut2017_snow.csv", package = "autocart")))
snow <- snow[1:40, c("yr50", "LONGITUDE", "LATITUDE", "ELEVATION", "MCMT", "PPTWT", "HUC")]
snow <- sp::SpatialPointsDataFrame(coords = snow[, c("LONGITUDE", "LATITUDE")],
                                   data = snow)

# Create an autocart model
snow_model <- autocart(yr50 ~ ELEVATION + MCMT + PPTWT + HUC, data = snow,
                       alpha = 0.30, beta = 0)

# Predict in autocart
new_snow <- snow[1:10, ]
autocart_predictions <- predict(snow_model, new_snow)
}
