% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autocart.R
\name{rmae}
\alias{rmae}
\title{Relative mean absolute error}
\usage{
rmae(pred, obs, na.rm = TRUE)
}
\arguments{
\item{pred}{The vector of predictions}

\item{obs}{The actual observed values}

\item{na.rm}{Should NA values be taken out of the vectors?}
}
\value{
The relative mean average error of the two vectors.
}
\description{
Relative mean absolute error
}
\examples{
# Create two vectors, add some noise, and evaluate the RMAE.
firstVec <- 1:10
secondVec <- 1:10 + rnorm(10)
rmae(firstVec, secondVec)
}
