% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialNodes.R
\name{spatialNodes}
\alias{spatialNodes}
\title{Using an autocart model, use the terminal nodes to form a spatial process that uses inverse
distance weighting to interpolate. The prediction for the new data that is passed in is formed
by making a prediction to assign it to a group. Next, the residual for the new prediction is
formed by inverse distance weighting the residual for the other points that are a part of that geometry.}
\usage{
spatialNodes(
  autocartModel,
  newdata,
  newdataCoords,
  method = "idw",
  distpower = 2,
  distpowerRange = c(0, 2),
  decideByGC = FALSE
)
}
\arguments{
\item{autocartModel}{an autocart model returned from the \code{autocart} function.}

\item{newdata}{a dataframe that contains the same predictors that were used to form the tree.}

\item{newdataCoords}{a matrix of coordinates for all the predictors contained in \code{newdata}}

\item{method}{The type of interpolation to use. Options are "idw" for inverse distance weighting and "tps" for thin-plate splines.}

\item{distpower}{the power to use if you would like to use something other than straight inverse distance, such as inverse distance squared.}

\item{distpowerRange}{A range of distpower to use. This is an adaptive inverse distance weighting method that linearly matches measures of spatial autocorrelation measured by Moran I to the range mentioned in distpower.}

\item{decideByGC}{When determining if a spatial process should be ran at a terminal node, should we use the Geary C statistic instead of Moran I?}
}
\value{
a prediction for the observations that are represented by \code{newdata} and \code{newdataCoords}
}
\description{
Using an autocart model, use the terminal nodes to form a spatial process that uses inverse
distance weighting to interpolate. The prediction for the new data that is passed in is formed
by making a prediction to assign it to a group. Next, the residual for the new prediction is
formed by inverse distance weighting the residual for the other points that are a part of that geometry.
}
